Add files via upload
Emotion detection is one of the most researched topics in the modern-day machine learning arena . The ability to accurately detect and identify an emotion opens up numerous doors for Advanced Human Computer Interaction. Detection of emotions can be done through human speech, body posture and facial expression. The aim of this project is to detect up to five distinct facial emotions in real time. This project runs on top of a Convolutional Neural Network (CNN) that is built with the help of Keras whose backend is TensorFlow in Python. The facial emotions that can be detected and classified by this system are Happy, Sad, Anger, Surprise and Neutral.